feat: containerize libs for microservices

+++ assets/scripts/build_image.sh
@@ -2,23 +2,55 @@
 
 # Args:
 # $1: the library for which you would like to build a container
-# --run ($2): whether to run the library container after build
-# --prune ($3): whether to prune docker system before building anything.
-# Run at root of repo!
+# --run: whether to run the library container after build
+# --prune: whether to prune docker system before building anything.
+# Run at the root of the repo!
+
+set -e
+
+# Default values for flags
+# run_=false
+# prune=false
+#
+# # Parse arguments
+# lib=""
+#
+# while [[ $# -gt 0 ]]; do
+#   case "$1" in
+#     --run)
+#       run_=true
+#       shift # Remove --run from processing
+#       ;;
+#     --prune)
+#       prune=true
+#       shift # Remove --prune from processing
+#       ;;
+#     *)
+#       if [[ -z "$lib" ]]; then
+#         lib="$1"
+#       else
+#         echo "Unknown argument: $1"
+#         exit 1
+#       fi
+#       shift # Remove the argument from processing
+#       ;;
+#   esac
+# done
 
 lib="$1"
-run_="$2" # --run
-prune="$3"  # --prune
+version="$2"
+prune="$3"
 
+if [[ -z "$lib" ]]; then
+  echo "You must specify the library to build."
+  exit 1
+fi
 
+# Prune the Docker system if --prune flag is set
 if [ "$prune" == "--prune" ]; then
   yes | docker system prune -a
 fi
 
-cd bio_check/"$lib" || exit
-docker build -t ghcr.io/biosimulators/bio-check-"$lib" .
+# Build the Docker image
+docker build -f ./"$lib"/Dockerfile-"$lib" -t ghcr.io/biosimulators/bio-check-"$lib":"$version" ./"$lib"
 
-if [ "$run_" == "--run" ]; then
-  echo "Running container for $lib"
-  ../../assets/run_container.sh "$lib"
-fi
